+++ com.unity.cinemachine/Editor/Editors/CinemachineDeoccluderEditor.cs
@@ -1,44 +1,37 @@
+#if CINEMACHINE_PHYSICS
+
 using UnityEngine;
 using UnityEditor;
 using System.Collections.Generic;
+using UnityEditor.UIElements;
+using UnityEngine.UIElements;
 
-#if CINEMACHINE_PHYSICS
 namespace Cinemachine.Editor
 {
     [CustomEditor(typeof(CinemachineDeoccluder))]
     [CanEditMultipleObjects]
-    class CinemachineDeoccluderEditor : BaseEditor<CinemachineDeoccluder>
+    class CinemachineDeoccluderEditor : UnityEditor.Editor
     {
-        /// <summary>Get the property names to exclude in the inspector.</summary>
-        /// <param name="excluded">Add the names to this list</param>
-        protected override void GetExcludedPropertiesInInspector(List<string> excluded)
-        {
-            base.GetExcludedPropertiesInInspector(excluded);
-            if (!Target.AvoidObstacles)
-            {
-                excluded.Add(FieldPath(x => x.DistanceLimit));
-                excluded.Add(FieldPath(x => x.CameraRadius));
-                excluded.Add(FieldPath(x => x.Strategy));
-                excluded.Add(FieldPath(x => x.MaximumEffort));
-                excluded.Add(FieldPath(x => x.Damping));
-                excluded.Add(FieldPath(x => x.DampingWhenOccluded));
-                excluded.Add(FieldPath(x => x.SmoothingTime));
-            }
-            else if (Target.Strategy == CinemachineDeoccluder.ResolutionStrategy.PullCameraForward)
-            {
-                excluded.Add(FieldPath(x => x.MaximumEffort));
-            }
-        }
+        CinemachineDeoccluder Target => target as CinemachineDeoccluder;
 
-        public override void OnInspectorGUI()
-        {
-            BeginInspector();
-            
-            CmPipelineComponentInspectorUtility.IMGUI_DrawMissingCmCameraHelpBox(this, Target.AvoidObstacles 
-                ? CmPipelineComponentInspectorUtility.RequiredTargets.LookAt 
-                : CmPipelineComponentInspectorUtility.RequiredTargets.None);
+        CmPipelineComponentInspectorUtility m_PipelineUtility;
 
-            DrawRemainingPropertiesInInspector();
+        void OnEnable() => m_PipelineUtility = new (this);
+        void OnDisable() => m_PipelineUtility.OnDisable();
+
+        public override VisualElement CreateInspectorGUI()
+        {
+            var ux = new VisualElement();
+
+            m_PipelineUtility.AddMissingCmCameraHelpBox(ux);
+            ux.Add(new PropertyField(serializedObject.FindProperty(() => Target.CollideAgainst)));
+            ux.Add(new PropertyField(serializedObject.FindProperty(() => Target.IgnoreTag)));
+            ux.Add(new PropertyField(serializedObject.FindProperty(() => Target.TransparentLayers)));
+            ux.Add(new PropertyField(serializedObject.FindProperty(() => Target.MinimumDistanceFromTarget)));
+            ux.Add(new PropertyField(serializedObject.FindProperty(() => Target.AvoidObstacles)));
+            ux.Add(new PropertyField(serializedObject.FindProperty(() => Target.ShotQualityEvaluation)));
+            m_PipelineUtility.UpdateState();
+            return ux;
         }
 
         [DrawGizmo(GizmoType.Active | GizmoType.Selected, typeof(CinemachineDeoccluder))]
@@ -49,14 +42,14 @@ namespace Cinemachine.Editor
             {
                 Color oldColor = Gizmos.color;
                 Vector3 pos = vcam.State.GetFinalPosition();
-                if (collider.AvoidObstacles && vcam.State.HasLookAt())
+                if (collider.AvoidObstacles.Enabled && vcam.State.HasLookAt())
                 {
                     Gizmos.color = CinemachineDeoccluderPrefs.CameraSphereColor.Value;
-                    if (collider.CameraRadius > 0)
-                        Gizmos.DrawWireSphere(pos, collider.CameraRadius);
+                    if (collider.AvoidObstacles.CameraRadius > 0)
+                        Gizmos.DrawWireSphere(pos, collider.AvoidObstacles.CameraRadius);
 
                     Vector3 forwardFeelerVector = (vcam.State.ReferenceLookAt - pos).normalized;
-                    float distance = collider.DistanceLimit;
+                    float distance = collider.AvoidObstacles.DistanceLimit;
                     Gizmos.DrawLine(pos, pos + forwardFeelerVector * distance);
 
                     // Show the avoidance path, for debugging